pyproject.toml
[project]
name = "ai-news-brief-personal"
version = "0.1.0"
description = "Add your description here"
requires-python = "==3.13.*"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "flask>=3.1.2",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.5",
    "requests>=2.32.5",
]
